make api protected

+++ src/module/Blog/blogsAdmin.controller.js
@@ -7,6 +7,13 @@ const cloudinary = require("cloudinary").v2;
 
 exports.createBlog = async (req, res) => {
   try {
+    const { email: userEmail } = req.user;
+    if (!userEmail) {
+      return res.status(400).json({
+        status: false,
+        message: "User not found",
+      });
+    }
     const { blogTitle, blogDescription } = req.body;
 
     if (!blogTitle || !blogDescription) {
@@ -115,6 +122,13 @@ exports.getSingleBlog = async (req, res) => {
 
 exports.updateBlog = async (req, res) => {
   try {
+    const { email: userEmail } = req.user;
+    if (!userEmail) {
+      return res.status(400).json({
+        status: false,
+        message: "User not found",
+      });
+    }
     const id = req.params.id;
     const { blogTitle, blogDescription } = req.body;
 
@@ -161,6 +175,13 @@ exports.updateBlog = async (req, res) => {
 
 exports.deleteBlog = async (req, res) => {
   try {
+    const { email: userEmail } = req.user;
+    if (!userEmail) {
+      return res.status(400).json({
+        status: false,
+        message: "User not found",
+      });
+    }
     const id = req.params.id;
     const blog = await blogsAdmin.findByIdAndDelete(id);
     if (!blog) {
